+++ diesel_compile_tests/tests/compile-fail/custom_returning_requires_nonaggregate.rs
@@ -13,7 +13,8 @@ table! {
     }
 }
 
-#[insertable_into(users)]
+#[derive(Insertable)]
+#[table_name="users"]
 pub struct NewUser {
     name: String,
 }