+++ diesel_compile_tests/tests/compile-fail/custom_returning_requires_selectable_expression.rs
@@ -19,7 +19,8 @@ table! {
     }
 }
 
-#[insertable_into(users)]
+#[derive(Insertable)]
+#[table_name="users"]
 pub struct NewUser {
     name: String,
 }