+++ diesel_compile_tests/tests/compile-fail/sqlite_upsert_cannot_be_used_on_pg.rs
@@ -13,7 +13,8 @@ table! {
     }
 }
 
-#[insertable_into(users)]
+#[derive(Insertable)]
+#[table_name="users"]
 struct User {
     id: i32,
 }